refactor: 파생 상태 관련 메서드 파일 네이밍, 디렉터리 변경
