+++ src/components/FilterContainer/SortFilter/index.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 import Filter from '../../common/Filter';
-import { RestaurantSortOption } from '../../../RestaurantUtils';
+import { RestaurantSortOption } from '../../../helpers/RestaurantHelper';
 
 type SortFilterProps = {
   onChange: (sortOption: RestaurantSortOption) => void;
